pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Milenakindleindarosa", email = "milena.kindlein.da.rosa.ifc"},
]
dependencies = ["django>=5.2.6", "djangorestframework>=3.16.1", "djangorestframework-simplejwt>=5.5.1", "django-cors-headers>=4.9.0", "rest-framework-simplejwt>=0.0.2"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
